--- a/go/hack/runtime.go
+++ b/go/hack/runtime.go
@@ -52,8 +52,11 @@ func RuntimeAllocSize(size int64) int64 {
 	return int64(roundupsize(uintptr(size)))
 }
 
-//go:linkname ParseFloatPrefix strconv.parseFloatPrefix
-func ParseFloatPrefix(s string, bitSize int) (float64, int, error)
+//go:linkname Atof64 strconv.atof64
+func Atof64(s string) (float64, int, error)
+
+//go:linkname Atof32 strconv.atof32
+func Atof32(s string) (float32, int, error)
 
 //go:linkname FastRand runtime.fastrand
 func FastRand() uint32