--- a/go/mysql/fastparse/fastparse.go
+++ b/go/mysql/fastparse/fastparse.go
@@ -234,7 +234,7 @@ func ParseFloat64(s string) (float64, error) {
 	// We only care to parse as many of the initial float characters of the
 	// string as possible. This functionality is implemented in the `strconv` package
 	// of the standard library, but not exposed, so we hook into it.
-	val, l, err := hack.ParseFloatPrefix(s[ws:], 64)
+	val, l, err := hack.Atof64(s[ws:])
 	for l < len(s[ws:]) {
 		if !isSpace(s[ws+uint(l)]) {
 			break