--- a/go/vt/vtgate/engine/cached_size.go
+++ b/go/vt/vtgate/engine/cached_size.go
@@ -579,7 +579,7 @@ func (cached *MemorySort) CachedSize(alloc bool) int64 {
 	if cc, ok := cached.UpperLimit.(cachedObject); ok {
 		size += cc.CachedSize(true)
 	}
-	// field OrderBy []vitess.io/vitess/go/vt/vtgate/evalengine.OrderByParams
+	// field OrderBy vitess.io/vitess/go/vt/vtgate/evalengine.Comparison
 	{
 		size += hack.RuntimeAllocSize(int64(cap(cached.OrderBy)) * int64(27))
 	}
@@ -606,7 +606,7 @@ func (cached *MergeSort) CachedSize(alloc bool) int64 {
 			}
 		}
 	}
-	// field OrderBy []vitess.io/vitess/go/vt/vtgate/evalengine.OrderByParams
+	// field OrderBy vitess.io/vitess/go/vt/vtgate/evalengine.Comparison
 	{
 		size += hack.RuntimeAllocSize(int64(cap(cached.OrderBy)) * int64(27))
 	}
@@ -799,7 +799,7 @@ func (cached *Route) CachedSize(alloc bool) int64 {
 	size += hack.RuntimeAllocSize(int64(len(cached.TableName)))
 	// field FieldQuery string
 	size += hack.RuntimeAllocSize(int64(len(cached.FieldQuery)))
-	// field OrderBy []vitess.io/vitess/go/vt/vtgate/evalengine.OrderByParams
+	// field OrderBy vitess.io/vitess/go/vt/vtgate/evalengine.Comparison
 	{
 		size += hack.RuntimeAllocSize(int64(cap(cached.OrderBy)) * int64(27))
 	}