… go/mysql/collations/colldata/uca_test.go
… go/vt/vtgate/engine/comparer_test.go
… go/vt/vtgate/engine/fake_vcursor_test.go
… go/vt/vtgate/engine/memory_sort_test.go
… go/vt/vtgate/engine/merge_sort.go
… go/vt/vtgate/engine/merge_sort_test.go
… go/vt/vtgate/engine/ordered_aggregate.go
… go/vt/vtgate/engine/route.go
… go/vt/vtgate/engine/route_test.go
… go/vt/vtgate/evalengine/api_compare.go
… go/vt/vtgate/evalengine/api_compare_test.go
… go/vt/vtgate/evalengine/weights.go
… go/vt/vtgate/evalengine/weights_test.go
… go/vt/vtgate/executor_framework_test.go
… go/vt/vtgate/executor_select_test.go
… go/vt/vtgate/planbuilder/operator_transformers.go
… go/vt/vttablet/tabletmanager/vdiff/utils.go
… go/vt/wrangler/vdiff.go