@vmg 2023-11-07
This was actually huge in `arewefast`.

![image](https://github.com/vitessio/vitess/assets/42793/2f8f8a81-1549-4b4c-8964-2be1219dbf28)
![image](https://github.com/vitessio/vitess/assets/42793/37b64fa3-7bba-46f3-bceb-184a41164af9)